Added scoring functions into services/patient_scoring.py from resources.py
